Use i18n fallbacks when looking up localized paths
In #853, the usage of i18n-fallbacks was introduced. For undocumented reasons, the fallbacks are ignored for localized paths, though.

This commit removes this limitation, adds a test to make sure, that it works as expected, and extends fixtures to make sure, that the other tests work with the changed behaviour.