Keeps full file path for chained templates
When handling chained templates, middleman used to remove the path and kept only the file's basename. Template engines which support imports (e.g. less or sass), use the file's path to find files to include. Since the path was removed, imports were not working for chained templates.

This commit changes middleman to keep the full path to the template file and adds a test case, which illustrates the desired behavior.